Add files via upload
Description:
Mtcars, an automobile company in Chambersburg, United States, has recorded the production of its cars within a dataset. The company is coming up with a new model based on the feedback given by its customers. It has to explore the current dataset to derive further insights from it.

Objective:
Import the dataset, explore for dimensionality, and type and average value of the horsepower across all the cars.
Also, identify a few of mostly correlated features, which would help in modification.